reorganized Visualizations/Comparison/Chart files
