+++ src/components/Visualizations/TypeOfRequest.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'proptypes';
 import { connect } from 'react-redux';
 import { REQUEST_TYPES } from '@components/common/CONSTANTS';
-import PieChart from './PieChart';
+import PieChart from '@components/Chart/PieChart';
 
 const TypeOfRequest = ({
   typeCounts,
